==> src/data_pipeline/05-feature_engg.py <==
import pandas as pd
import os

df = pd.read_csv("../../datasets/validation/validated_df.csv")

df['can_fly'] = (df['airborne'] == 1) & (df['feathers'] == 1)
df['can_swim'] = (df['aquatic'] == 1) & (df['fins'] == 1)
df['is_domestic_pet'] = (df['domestic'] == 1) & (df['catsize'] == 1)

# convert False, True to numerical 0, 1
df['can_fly'] = df['can_fly'].astype(int)
df['can_swim'] = df['can_swim'].astype(int)
df['is_domestic_pet'] = df['is_domestic_pet'].astype(int)

# save feature_engg dataset
output_dir = "../../datasets/feature_engg"
os.makedirs(output_dir, exist_ok=True)

df.to_csv(f"{output_dir}/feature_df.csv", index_label=False)
print(df.head(3))
